+++ bio_embeddings/extract/bindEmbed21DL/__init__.py
@@ -1,0 +1,5 @@
+from bio_embeddings.extract.bindEmbed21DL.bindEmbed21DL_annotation_extractor import (
+    BindEmbed21DLAnnotationExtractor,
+)
+
+__all__ = ["BindEmbed21DLAnnotationExtractor"]