+++ bio_embeddings/extract/light_attention/__init__.py
@@ -1,1 +1,6 @@
-from bio_embeddings.extract.light_attention.LightAttentionAnnotationExtractor import LightAttentionAnnotationExtractor
+from bio_embeddings.extract.light_attention.light_attention_annotation_extractor import (
+    LightAttentionAnnotationExtractor,
+)
+from bio_embeddings.extract.light_attention.light_attention_model import LightAttention
+
+__all__ = ["LightAttentionAnnotationExtractor", "LightAttention"]