+++ bio_embeddings/extract/pipeline.py
@@ -1,22 +1,25 @@
 import logging
-import h5py
-import numpy as np
-
 from copy import deepcopy
 from typing import Dict, Any, List
+
+import h5py
+import numpy as np
 from Bio.Seq import Seq
 from pandas import read_csv, DataFrame, concat as concatenate_dataframe
 from sklearn.metrics import pairwise_distances as _pairwise_distances
 
 from bio_embeddings.extract.basic import BasicAnnotationExtractor
-from bio_embeddings.extract.light_attention.LightAttentionAnnotationExtractor import LightAttentionAnnotationExtractor
+from bio_embeddings.extract.light_attention.light_attention_annotation_extractor import \
+    LightAttentionAnnotationExtractor
+from bio_embeddings.extract.prott5cons import ProtT5consAnnotationExtractor
+from bio_embeddings.extract.bindEmbed21DL.bindEmbed21DL_annotation_extractor import BindEmbed21DLAnnotationExtractor
 from bio_embeddings.extract.unsupervised_utilities import get_k_nearest_neighbours
-from bio_embeddings.utilities.remote_file_retriever import get_model_file
+from bio_embeddings.utilities import get_model_file
+from bio_embeddings.utilities.exceptions import InvalidParameterError, UnrecognizedEmbeddingError, \
+    InvalidAnnotationFileError
 from bio_embeddings.utilities.filemanagers import get_file_manager
 from bio_embeddings.utilities.helpers import check_required, read_fasta, convert_list_of_enum_to_string, \
     write_fasta_file, read_mapping_file
-from bio_embeddings.utilities.exceptions import InvalidParameterError, UnrecognizedEmbeddingError, \
-    InvalidAnnotationFileError
 
 logger = logging.getLogger(__name__)
 
@@ -195,8 +198,10 @@ def seqvec_from_publication(**kwargs) -> Dict[str, Any]:
 def bert_from_publication(**kwargs) -> Dict[str, Any]:
     return predict_annotations_using_basic_models("bert_from_publication", **kwargs)
 
+
 def t5_xl_u50_from_publication(**kwargs) -> Dict[str, Any]:
     return predict_annotations_using_basic_models("t5_xl_u50_from_publication", **kwargs)
+
 
 def la_prott5(**kwargs) -> Dict[str, Any]:
     return light_attention('la_prott5', **kwargs)
@@ -206,7 +211,115 @@ def la_protbert(**kwargs) -> Dict[str, Any]:
     return light_attention('la_protbert', **kwargs)
 
 
-def light_attention(model, **kwargs) -> Dict[str, Any]:
+def prott5cons(model: str, **kwargs) -> Dict[str, Any]:
+    """
+    Protocol extracts conservation from "embeddings_file".
+    Embeddings can only be generated with ProtT5-XL-U50.
+
+    :param model: "t5_xl_u50_conservation". Used to download files
+    """
+
+    check_required(kwargs, ['embeddings_file', 'mapping_file', 'remapped_sequences_file'])
+    result_kwargs = deepcopy(kwargs)
+    file_manager = get_file_manager(**kwargs)
+
+    # Download necessary files if needed
+    for file in ProtT5consAnnotationExtractor.necessary_files:
+        if not result_kwargs.get(file):
+            result_kwargs[file] = get_model_file(model=model, file=file)
+
+    annotation_extractor = ProtT5consAnnotationExtractor(**result_kwargs)
+
+    # mapping file will be needed for protein-wide annotations
+    mapping_file = read_mapping_file(result_kwargs["mapping_file"])
+
+    # Try to create final files (if this fails, now is better than later
+    conservation_predictions_file_path = file_manager.create_file(result_kwargs.get('prefix'),
+                                                                  result_kwargs.get('stage_name'),
+                                                                  'conservation_predictions_file',
+                                                                  extension='.fasta')
+    result_kwargs['conservation_predictions_file'] = conservation_predictions_file_path
+    cons_sequences = list()
+    with h5py.File(result_kwargs['embeddings_file'], 'r') as embedding_file:
+        for protein_sequence in read_fasta(result_kwargs['remapped_sequences_file']):
+            embedding = np.array(embedding_file[protein_sequence.id])
+
+            annotations = annotation_extractor.get_conservation(embedding)
+            cons_sequence = deepcopy(protein_sequence)
+            cons_sequence.seq = Seq(convert_list_of_enum_to_string(annotations.conservation))
+            cons_sequences.append(cons_sequence)
+
+    # Write files
+    write_fasta_file(cons_sequences, conservation_predictions_file_path)
+    return result_kwargs
+
+
+def bindembed21dl(**kwargs) -> Dict[str, Any]:
+    """
+    Protocol extracts binding residues from "embeddings_file".
+    Results guaranteed only with ProtT5-XL-U50 embeddings.
+
+    :return:
+    """
+
+    check_required(kwargs, ['embeddings_file', 'mapping_file', 'remapped_sequences_file'])
+    result_kwargs = deepcopy(kwargs)
+    file_manager = get_file_manager(**kwargs)
+
+    # Download necessary files if needed
+    for file in BindEmbed21DLAnnotationExtractor.necessary_files:
+        if not result_kwargs.get(file):
+            result_kwargs[file] = get_model_file(model="bindembed21dl", file=file)
+
+    annotation_extractor = BindEmbed21DLAnnotationExtractor(**result_kwargs)
+
+    # Try to create final files (if this fails, now is better than later
+    metal_binding_predictions_file_path = file_manager.create_file(result_kwargs.get('prefix'),
+                                                                   result_kwargs.get('stage_name'),
+                                                                   'metal_binding_predictions_file',
+                                                                   extension='.fasta')
+    result_kwargs['metal_binding_predictions_file'] = metal_binding_predictions_file_path
+    nuc_binding_predictions_file_path = file_manager.create_file(result_kwargs.get('prefix'),
+                                                                 result_kwargs.get('stage_name'),
+                                                                 'nucleic_acid_binding_predictions_file',
+                                                                 extension='.fasta')
+    result_kwargs['binding_residue_predictions_file'] = nuc_binding_predictions_file_path
+    small_binding_predictions_file_path = file_manager.create_file(result_kwargs.get('prefix'),
+                                                                   result_kwargs.get('stage_name'),
+                                                                   'small_molecule_binding_predictions_file',
+                                                                   extension='.fasta')
+    result_kwargs['binding_residue_predictions_file'] = small_binding_predictions_file_path
+
+    metal_sequences = list()
+    nuc_sequences = list()
+    small_sequences = list()
+
+    with h5py.File(result_kwargs['embeddings_file'], 'r') as embedding_file:
+        for protein_sequence in read_fasta(result_kwargs['remapped_sequences_file']):
+            embedding = np.array(embedding_file[protein_sequence.id])
+
+            annotations = annotation_extractor.get_binding_residues(embedding)
+            metal_sequence = deepcopy(protein_sequence)
+            nuc_sequence = deepcopy(protein_sequence)
+            small_sequence = deepcopy(protein_sequence)
+
+            metal_sequence.seq = Seq(convert_list_of_enum_to_string(annotations.metal_ion))
+            nuc_sequence.seq = Seq(convert_list_of_enum_to_string(annotations.nucleic_acids))
+            small_sequence.seq = Seq(convert_list_of_enum_to_string(annotations.small_molecules))
+
+            metal_sequences.append(metal_sequence)
+            nuc_sequences.append(nuc_sequence)
+            small_sequences.append(small_sequence)
+
+    # Write files
+    write_fasta_file(metal_sequences, metal_binding_predictions_file_path)
+    write_fasta_file(nuc_sequences, nuc_binding_predictions_file_path)
+    write_fasta_file(small_sequences, small_binding_predictions_file_path)
+
+    return result_kwargs
+
+
+def light_attention(model: str, **kwargs) -> Dict[str, Any]:
     """
     Protocol extracts subcellular locationfrom "embeddings_file".
     Embeddings can be generated with ProtBert.
@@ -218,12 +331,7 @@ def light_attention(model, **kwargs) -> Dict[str, Any]:
     result_kwargs = deepcopy(kwargs)
     file_manager = get_file_manager(**kwargs)
 
-    # Download necessary files if needed
-    for file in LightAttentionAnnotationExtractor.necessary_files:
-        if not result_kwargs.get(file):
-            result_kwargs[file] = get_model_file(model=model, file=file)
-
-    annotation_extractor = LightAttentionAnnotationExtractor(**result_kwargs)
+    annotation_extractor = LightAttentionAnnotationExtractor(model, **result_kwargs)
 
     # mapping file will be needed for protein-wide annotations
     mapping_file = read_mapping_file(result_kwargs["mapping_file"])
@@ -251,7 +359,7 @@ def light_attention(model, **kwargs) -> Dict[str, Any]:
     return result_kwargs
 
 
-def predict_annotations_using_basic_models(model, **kwargs) -> Dict[str, Any]:
+def predict_annotations_using_basic_models(model: str, **kwargs) -> Dict[str, Any]:
     """
     Protocol extracts secondary structure (DSSP3 and DSSP8), disorder, subcellular location and membrane boundness
     from "embeddings_file". Embeddings can either be generated with SeqVec or ProtBert.
@@ -264,11 +372,6 @@ def predict_annotations_using_basic_models(model, **kwargs) -> Dict[str, Any]:
     check_required(kwargs, ['embeddings_file', 'mapping_file', 'remapped_sequences_file'])
     result_kwargs = deepcopy(kwargs)
     file_manager = get_file_manager(**kwargs)
-
-    # Download the checkpoint files if needed
-    for file in BasicAnnotationExtractor.necessary_files:
-        if not result_kwargs.get(file):
-            result_kwargs[file] = get_model_file(model=f'{model}_annotations_extractors', file=file)
 
     annotation_extractor = BasicAnnotationExtractor(model, **result_kwargs)
 
@@ -386,6 +489,8 @@ PROTOCOLS = {
     "t5_xl_u50_from_publication": t5_xl_u50_from_publication,
     "la_prott5": la_prott5,
     "la_protbert": la_protbert,
+    "prott5cons": prott5cons,
+    "bindembed21dl": bindembed21dl,
     "unsupervised": unsupervised
 }