+++ bio_embeddings/extract/prott5cons/__init__.py
@@ -1,0 +1,6 @@
+from bio_embeddings.extract.prott5cons.prot_t5_cons_annotation_extractor import (
+    ProtT5consAnnotationExtractor,
+    ConservationCNN,
+)
+
+__all__ = ["ConservationCNN", "ProtT5consAnnotationExtractor"]