+++ webserver/utilities/parsers.py
@@ -30,7 +30,7 @@ class SecondaryStructure(Enum):
         }.get(self)
 
     @staticmethod
-    def isAAFeature():
+    def is_aa_feature():
         return True
 
 
@@ -47,7 +47,7 @@ class Disorder(Enum):
         }.get(self)
 
     @staticmethod
-    def isAAFeature():
+    def is_aa_feature():
         return True